okestra/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.35.0
django-appconf==1.0.2
django-authtools==1.6.0
django-bootstrap4==0.0.6
django-compressor==2.2
django-libsass==0.7
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
idna==2.6
libsass==0.14.5
oauthlib==2.0.7
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dotenv==0.8.2
python3-openid==3.1.0
pytz==2018.4
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
urllib3==1.22
